requirements.txt
APScheduler==3.6.3
attrs==22.2.0
autopep8==2.0.1
cachetools==4.2.2
certifi==2022.12.7
coverage==7.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pycodestyle==2.10.0
pytest==7.2.2
pytest-cov==4.0.0
python-dotenv==0.21.1
python-telegram-bot==13.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tomli==2.0.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
